Convert history-based router to use hashes.
The strategy is that we wire location.set to be called on `hashchange`
rather than on popstate and the synthetic pushstate events.

This also uses an exclamation point after the first hash, so
`/some/path/here` is denoted `#!/some/path/here`. I mostly have
in mind that I want to be able to have anchors on a path, like
`/some/path#anchor`, and `#!/some/path#anchor` looks slightly less
weird than `#/some/path#anchor`.

Currently the hash prefix isn't customizable.

Also, I removed Redirect because it was just going to be a one-liner
that set window.location.hash, but that wouldn't have had the same
behavior as history router's Redirect, which calls `replaceState` and
so avoids adding an extra browser history entry.